% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca.R
\name{algorithm_normalization}
\alias{algorithm_normalization}
\title{FQPCA normalization}
\usage{
algorithm_normalization(loadings, scores)
}
\arguments{
\item{loadings}{Matrix of loadings}

\item{scores}{Matrix of scores}
}
\value{
The normalized matrices of loadings and scores and the normalization matrix.
}
\description{
Performs the normalization of matrices of loadings and scores in order to ensure the solution is unique.
}
