% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_tensor_matrix_no_intercept}
\alias{build_tensor_matrix_no_intercept}
\title{Build Tensor Design Matrix}
\usage{
build_tensor_matrix_no_intercept(scores, Y.mask, spline.basis)
}
\arguments{
\item{scores}{A matrix of functional Principal Component Analysis (fPCA) scores for each subject.}

\item{Y.mask}{A logical mask matrix of dimensions (N x T) indicating which time points are observed (\code{TRUE}) or missing (\code{FALSE}) for each subject.}

\item{spline.basis}{The evaluation matrix of the B-spline basis functions at the time points.}
}
\value{
A matrix containing the stacked tensor products of the spline bases for all valid observations.
}
\description{
Inner function to construct the tensor product design matrix required for functional data modeling. It efficiently scales the B-spline basis evaluations by the subject-specific functional principal component scores across all observed time points.
}
