% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-checks.R
\name{check_flag}
\alias{check_flag}
\title{Check that a parameter is a single Boolean}
\usage{
check_flag(x, name)
}
\arguments{
\item{x}{The value to check.}

\item{name}{The parameter name used in the error message.}
}
\value{
No return
}
\description{
Check that a parameter is a single Boolean
}
