% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosqr_fqpca.R
\name{check_fosqr_fqpca_params}
\alias{check_fosqr_fqpca_params}
\title{Check input parameters}
\usage{
check_fosqr_fqpca_params(
  npc,
  quantile.value,
  periodic,
  splines.df,
  splines.method,
  tol,
  max.iters,
  verbose,
  estimate.variance,
  n.bootstrap,
  seed
)
}
\arguments{
\item{npc}{The number of estimated components.}

\item{quantile.value}{The quantile considered.}

\item{periodic}{Boolean indicating if the data is expected to be periodic (start coincides with end) or not.}

\item{splines.df}{Degrees of freedom for the splines.}

\item{splines.method}{Method used in the resolution of the splines quantile regression model. It currently accepts the methods \code{c('conquer', 'quantreg')}.}

\item{tol}{Tolerance on the convergence of the algorithm.}

\item{max.iters}{Maximum number of iterations.}

\item{verbose}{Boolean indicating the verbosity.}

\item{estimate.variance}{Booleand indicating if variance for the FOSQR loadings should be estimated}

\item{n.bootstrap}{Number of bootstrap resamplings used in the estimation of the variance. Only in effect if estimate.variance=TRUE}

\item{seed}{Seed for the random generator number.}
}
\value{
No return
}
\description{
Check input parameters
}
