% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosqr.R
\name{check_fosqr_params}
\alias{check_fosqr_params}
\title{Check input parameters}
\usage{
check_fosqr_params(
  quantile.value,
  periodic,
  splines.df,
  splines.method,
  verbose,
  seed
)
}
\arguments{
\item{quantile.value}{The quantile considered.}

\item{periodic}{Boolean indicating if the data is expected to be periodic (start coincides with end) or not.}

\item{splines.df}{Degrees of freedom for the splines.}

\item{splines.method}{Method used in the resolution of the splines quantile regression model. It currently accepts the methods \code{c('conquer', 'quantreg')}.}

\item{verbose}{Boolean indicating the verbosity.}

\item{seed}{Seed for the random generator number.}
}
\value{
No return
}
\description{
Check input parameters for fosqr
}
