% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca.R
\name{check_fqpca_params}
\alias{check_fqpca_params}
\title{Check input parameters}
\usage{
check_fqpca_params(
  npc,
  quantile.value,
  periodic,
  splines.df,
  splines.method,
  penalized,
  lambda.ridge,
  tol,
  n.iters,
  verbose,
  seed,
  parallelized.scores,
  num.cores
)
}
\arguments{
\item{npc}{The number of estimated components.}

\item{quantile.value}{The quantile considered.}

\item{periodic}{Boolean indicating if the data is expected to be periodic (start coincides with end) or not.}

\item{splines.df}{Degrees of freedom for the splines.}

\item{splines.method}{Method used in the resolution of the splines quantile regression model. It currently accepts the methods \code{c('conquer', 'quantreg')}.}

\item{penalized}{Boolean indicating if the smoothness should be controlled using a second derivative penalty. This functionality is experimental and is much slower than the control of the smoothness using the degrees of freedom.}

\item{lambda.ridge}{Hyper parameter controlling the penalization on the second derivative of the splines. It has effect only with \code{penalized=TRUE}.}

\item{tol}{Tolerance on the convergence of the algorithm.}

\item{n.iters}{Maximum number of iterations.}

\item{verbose}{Boolean indicating the verbosity.}

\item{seed}{Seed for the random generator number.}

\item{parallelized.scores}{Should the scores be computed in parallel? Experimantal component.}

\item{num.cores}{Number of cores to use in parallelized executions.}
}
\value{
No return
}
\description{
Check input parameters
}
