% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca.R
\name{check_input_data}
\alias{check_input_data}
\title{Check input data}
\usage{
check_input_data(data, colname)
}
\arguments{
\item{data}{An \eqn{(N \times T)} matrix, a tf object from the tidyfun package or a data.frame containing the functional data as a tf column.}

\item{colname}{The name of the column containing the functional data. Use only if data is a dataframe and colname is a column in the dataframe.}
}
\value{
an unnamed matrix
}
\description{
Check input data
}
