% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-checks.R
\name{check_iters}
\alias{check_iters}
\title{Check that a maximum-iterations parameter is an integer >= 2}
\usage{
check_iters(x, name)
}
\arguments{
\item{x}{The value to check.}

\item{name}{The parameter name used in the error message.}
}
\value{
No return
}
\description{
Check that a maximum-iterations parameter is an integer >= 2
}
