% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-quantile.R
\name{check_loss}
\alias{check_loss}
\title{Quantile check (pinball) loss}
\usage{
check_loss(x, quantile.value = 0.5)
}
\arguments{
\item{x}{The vector.}

\item{quantile.value}{The quantile considered. Defaults to 0.5.}
}
\value{
Objective value function for a quantile regression model.
}
\description{
Evaluates the quantile check loss elementwise.
}
