% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-checks.R
\name{check_quantile_value}
\alias{check_quantile_value}
\title{Check that the quantile value lies in (0, 1)}
\usage{
check_quantile_value(x)
}
\arguments{
\item{x}{The value to check.}
}
\value{
No return
}
\description{
Check that the quantile value lies in (0, 1)
}
