% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-checks.R
\name{check_seed}
\alias{check_seed}
\title{Check that the seed is a positive integer or NULL}
\usage{
check_seed(seed)
}
\arguments{
\item{seed}{The value to check.}
}
\value{
No return
}
\description{
Check that the seed is a positive integer or NULL
}
