% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-checks.R
\name{check_splines_method}
\alias{check_splines_method}
\title{Check that the splines method is supported}
\usage{
check_splines_method(x)
}
\arguments{
\item{x}{The value to check.}
}
\value{
No return
}
\description{
Check that the splines method is supported
}
