% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosqr_fqpca.R
\name{check_Y}
\alias{check_Y}
\title{Checks input data}
\usage{
check_Y(Y)
}
\arguments{
\item{Y}{An \eqn{(N \times T)} matrix or a tf object from the tidyfun package}
}
\value{
The prepared Y matrix
}
\description{
Checks that Y is valid and formats it.
}
