% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca.R
\name{compute_explained_variability}
\alias{compute_explained_variability}
\title{FQPCA explained variability}
\usage{
compute_explained_variability(scores)
}
\arguments{
\item{scores}{Matrix of scores}
}
\value{
The percentage of variability each component is explaining.
}
\description{
Computes the percentage of explained variability based on the variance of the scores matrix
}
