% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca_no_scaling.R
\name{compute_F_Lambda_no_scaling}
\alias{compute_F_Lambda_no_scaling}
\title{Compute loadings and scores}
\usage{
compute_F_Lambda_no_scaling(
  x,
  x_mask,
  quantile_value,
  basis_coef,
  spline_basis,
  spline_basis_i
)
}
\arguments{
\item{x}{The N by T matrix of observed time instants.}

\item{x_mask}{Mask matrix of the same dimensions asx indicating wihch observations in x are known.}

\item{quantile_value}{The quantile considered. Default is the median 0.5.}

\item{basis_coef}{The matrix of spline basis coefficients.}

\item{spline_basis}{The matrix of spline basis.}

\item{spline_basis_i}{A list containing, for each observation xi, the matrix of associated spline basis.}
}
\value{
The unnormalized matrices of loadings and scores, along with the mean and standard deviations of the scores (these last are used in the prediction function).
}
\description{
Given matrix x and spline information, computes the matrices of loadings and scores.
}
