% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca.R
\name{compute_loadings}
\alias{compute_loadings}
\title{Compute loadings (aka principal components)}
\usage{
compute_loadings(spline_basis, spline_coefficients)
}
\arguments{
\item{spline_basis}{The spline basis matrix.}

\item{spline_coefficients}{the matrix of spline coefficients}
}
\value{
The matrix of loadings
}
\description{
Inner function to compute the loadings of the fqpca methodology.
}
