% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca.R
\name{compute_objective_value}
\alias{compute_objective_value}
\title{Compute objective value function}
\usage{
compute_objective_value(x, quantile_value, scores, loadings)
}
\arguments{
\item{x}{The N by T matrix of observed time instants.}

\item{quantile_value}{The quantile considered.}

\item{scores}{The matrix of estimated scores}

\item{loadings}{The matrix of estimated loadings}
}
\value{
The objective value function
}
\description{
Inner function to compute the objective value
}
