% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosqr_fqpca.R
\name{compute_scores_fosr_fqpca}
\alias{compute_scores_fosr_fqpca}
\title{Compute scores}
\usage{
compute_scores_fosr_fqpca(
  Y,
  Y.mask,
  Y.fosr.hat,
  fqpca.loadings,
  quantile.value
)
}
\arguments{
\item{Y}{The \eqn{(N \times T)} matrix of observed time instants.}

\item{Y.mask}{Mask matrix of the same dimensions as Y indicating which observations in Y are known.}

\item{Y.fosr.hat}{fosqr fitted values}

\item{fqpca.loadings}{Matrix of loading coefficients.}

\item{quantile.value}{The quantile considered.}
}
\value{
The matrix of scores.
}
\description{
Inner function to compute the scores of the fosqr-fqpca methodology.
}
