% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca_lasso.R
\name{compute_scores_lasso}
\alias{compute_scores_lasso}
\title{Compute scores}
\usage{
compute_scores_lasso(x, x_mask, loadings, quantile_value, lambda)
}
\arguments{
\item{x}{The N by T matrix of observed time instants.}

\item{x_mask}{Mask matrix of the same dimensions as x indicating wich observations in x are known.}

\item{loadings}{Matrix of loading coefficients}

\item{quantile_value}{The quantile considered.}

\item{lambda}{A grid of hyper-parameter values for the lambda parameter in lasso penalized quantile regression.}
}
\value{
The matrix of scores
}
\description{
Inner function to compute the scores of the fqpca methodology.
}
