% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca.R
\name{compute_scores_sequential}
\alias{compute_scores_sequential}
\title{Compute scores}
\usage{
compute_scores_sequential(Y, Y.mask, loadings, quantile.value)
}
\arguments{
\item{Y}{The \eqn{(N \times T)} matrix of observed time instants.}

\item{Y.mask}{Mask matrix of the same dimensions as Y indicating which observations in Y are known.}

\item{loadings}{Matrix of loading coefficients.}

\item{quantile.value}{The quantile considered.}
}
\value{
The matrix of scores.
}
\description{
Inner function to compute the scores of the fqpca methodology.
}
