% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca.R
\name{compute_spline_coefficients_penalized}
\alias{compute_spline_coefficients_penalized}
\title{Compute spline coefficients}
\usage{
compute_spline_coefficients_penalized(
  Y,
  Y.mask,
  scores,
  spline.basis,
  intercept.spline.basis,
  quantile.value,
  lambda.ridge,
  weights
)
}
\arguments{
\item{Y}{\eqn{(N \times T)} matrix of observed time instants.}

\item{Y.mask}{Mask matrix of the same dimensions as Y indicating wich observations in Y are known.}

\item{scores}{Initial matrix of scores.}

\item{spline.basis}{The spline basis matrix.}

\item{intercept.spline.basis}{The spline basis matrix for the conquer intercept.}

\item{quantile.value}{The quantile considered.}

\item{lambda.ridge}{The hyper-parameter controlling the penalization on the splines. This parameter has no effect if method is 'quantreg'}

\item{weights}{Array of weights used in the penalization. It must have dimension equal to (npc+1) * ncol(spline.basis)-1. (The -1 comes from the presence of an intercept term)}
}
\value{
The matrix of spline coefficients splines coefficients.
}
\description{
Inner function to compute the spline coefficients of the fqpca methodology.
}
