% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca.R
\name{compute_spline_coefficients_unpenalized}
\alias{compute_spline_coefficients_unpenalized}
\title{Compute spline coefficients}
\usage{
compute_spline_coefficients_unpenalized(
  Y,
  Y.mask,
  scores,
  spline.basis,
  quantile.value,
  method
)
}
\arguments{
\item{Y}{\eqn{(N \times T)} matrix of observed time instants.}

\item{Y.mask}{Mask matrix of the same dimensions as Y indicating wich observations in Y are known.}

\item{scores}{Initial matrix of scores.}

\item{spline.basis}{The spline basis matrix.}

\item{quantile.value}{The quantile considered.}

\item{method}{Method used in the resolution of the quantile regression model. It currently accepts the methods \code{c('conquer', 'quantreg')}.}
}
\value{
The matrix of spline coefficients splines coefficients.
}
\description{
Inner function to compute the spline coefficients of the fqpca methodology.
}
