% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosqr_fqpca.R
\name{compute_var_correction}
\alias{compute_var_correction}
\title{Compute  correction variance associated to fosqr loadings}
\usage{
compute_var_correction(n.bootstrap, fqpca.scores, regressors, fqpca.loadings)
}
\arguments{
\item{n.bootstrap}{number of bootstraps}

\item{fqpca.scores}{matrix of fqpca scores}

\item{regressors}{matrix of covariates}

\item{fqpca.loadings}{matrix of fqpca loadings}
}
\value{
A matrix of variances associated to the fosqr loadings
}
\description{
Given a fosqr-fqpca splines coefficients model, estimates variances associated to fosqr loadings.
}
