% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosqr_fqpca.R
\name{compute_variance.fosqr_fqpca_object}
\alias{compute_variance.fosqr_fqpca_object}
\title{Compute variance}
\usage{
\method{compute_variance}{fosqr_fqpca_object}(object, pve = 0.95, n.bootstrap = 1000, ...)
}
\arguments{
\item{object}{An object output of the fosqr_fqpca function.}

\item{pve}{If smaller than 1, taken as percentage of explained variability used in variance estimation. If greater than 1, taken as number of components used in variance estimation.}

\item{n.bootstrap}{Number of bootstrap resamplings used in the estimation of the variance correction.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list containing the estimated total variance, analytical variance, correction variance, and the kernel covariance model.
}
\description{
S3 method for class 'fosqr_fqpca_object'. Given a fosqr_fqpca_object model, estimates the variance associated to the FOSQR loadings.
}
\examples{

n.obs = 150
n.time = 144
time.grid <- seq(0, 2 * pi, length.out = n.time)

# Generate regressor and principal component functions
b1 = sin(time.grid)
pc1 = sin(2 * time.grid)

# Generate covariates and scores
regressors =  5 * matrix(rnorm(n.obs), ncol=1)
scores = matrix(10 * rnorm(n.obs), ncol = 1)
epsilon.matrix <- 0.1 * matrix(rnorm(n.obs * n.time), nrow = n.obs)

# Build dataset
Y = 100 + regressors[, 1] \%o\% b1 + scores[, 1] \%o\% pc1 + epsilon.matrix

# Add missing observations
Y[sample(n.obs*n.time, as.integer(0.2*n.obs*n.time))] <- NA

results <- fosqr_fqpca(Y = Y, regressors=regressors, npc = 1, quantile.value = 0.5, seed=1)
fosqr_variances <- compute_variance(object = results)
}
