% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosqr.R
\name{compute_variance.fosqr_object}
\alias{compute_variance.fosqr_object}
\title{Compute variance}
\usage{
\method{compute_variance}{fosqr_object}(object, ...)
}
\arguments{
\item{object}{An object output of the fosqr function.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list containing the estimated analytical variance and the kernel covariance model.
}
\description{
S3 method for class 'fosqr_object'. Given a fosqr_object model, estimates the variance associated to the FOSQR loadings.
}
