% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosqr_fqpca.R
\name{compute_variance}
\alias{compute_variance}
\title{Compute variance}
\usage{
compute_variance(object, ...)
}
\arguments{
\item{object}{An object for which to compute variance.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Generic function to compute variance.
}
