% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosqr_fqpca.R
\name{compute_variance_fosqr}
\alias{compute_variance_fosqr}
\title{Compute variance associated to fosqr loadings}
\usage{
compute_variance_fosqr(object, Y, pve = 0.95, n.bootstrap = 1000)
}
\arguments{
\item{object}{An object output of the fosqr_fqpca function.}

\item{Y}{An \eqn{(N \times T)} matrix or a tf object from the tidyfun package.}

\item{pve}{If smaller than 1, taken as percentage of explained variability used in Yhat estimation. If greater than 1, taken as number of components  used in Yhat estimation.}

\item{n.bootstrap}{An integer number used in the bootstrapping part of the variance estimation.}
}
\value{
A list containing the two parts of the variance computation and the final variances.
}
\description{
S3 method for class 'fosqr_fqpca_object'. Given a fosqr_fqpca_object model, estimates variances associated to fosqr loadings.
}
\examples{
n.obs = 150
n.time = 144
time.grid <- seq(0, 2 * pi, length.out = n.time)

# Generate regressor and principal component functions
b1 = sin(time.grid)
pc1 = sin(2 * time.grid)

# Generate covariates and scores
regressors = 5 * matrix(rnorm(n.obs), ncol=1)
scores = matrix(10 * rnorm(n.obs), ncol = 1)
epsilon.matrix <- 0.1 * matrix(rnorm(n.obs * n.time), nrow = n.obs)

# Build dataset
Y = 100 + regressors[, 1] \%o\% b1 + scores[, 1] \%o\% pc1 + epsilon.matrix

# Add missing observations
Y[sample(n.obs*n.time, as.integer(0.2*n.obs*n.time))] <- NA

results <- fosqr_fqpca(Y = Y, regressors=regressors, npc = 1, quantile.value = 0.5, seed=1)
variances <- compute_variance_fosqr(object=results, Y=Y, pve=0.95)
}
