% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cross_validation_alpha}
\alias{cross_validation_alpha}
\title{CROSS VALIDATION}
\usage{
cross_validation_alpha(
  x,
  n_components = 1,
  quantile_value = 0.5,
  alpha_grid = c(0, 10^seq(-16, -5, by = 1)),
  n_folds = 3,
  criteria = "points",
  periodic = TRUE,
  splines_df = 30,
  tol = 0.001,
  n_iters = 50,
  solver = "SCS",
  penalized = TRUE,
  method = "br",
  verbose_fqpca = FALSE,
  verbose_cv = TRUE,
  seed = NULL
)
}
\arguments{
\item{x}{The N by T matrix of observed time instants}

\item{n_components}{The number of estimated components. Default is 2.}

\item{quantile_value}{The quantile considered. Default is the median 0.5.}

\item{alpha_grid}{An array containing the list of possible alpha values (these should be always positive numbers).}

\item{n_folds}{Number of folds to be used on the cross validation process. Default is 3}

\item{criteria}{Criteria used to divide the data. Valid values are 'rows', which considers the division based on full rows, or 'points', which considers the division based on points within the matrix. Default is points}

\item{periodic}{Boolean indicating if the data is expected to be periodic (start coincides with end) or not. Default is TRUE.}

\item{splines_df}{Degrees of freedom for the splines. It is recommended not to modify this parameter and control the smoothness via the hyperparameter alpha_ridge. Default is 30.}

\item{tol}{Tolerance on the convergence of the algorithm. Smaller values can spped up computation but may affect the quality of the estimations. Default is 1e-3.}

\item{n_iters}{Maximum number of iterations. Default is 30.}

\item{solver}{Solver to be used by the CVXR package in the resolution of the penalized quantile regression models. The speed of the methodology can vary depending on the solver used. There are free alternatives available like 'SCS', 'ECOS' or 'OSQP'. It is also possible to use licensed programs like 'MOSEK' or 'GUROBI'. Default is 'SCS'.}

\item{penalized}{Boolean indicating if the model should solve a penalized splines model or a non penelized splines model}

\item{method}{Method to be used by the quantreg package when solving quantile regression models. This parameter has no effect if penalized is TRUE (default)}

\item{verbose_fqpca}{Boolean indicating verbosity of the fqpca function. Default is FALSE.}

\item{verbose_cv}{Boolean indicating verbosity of the Cross validation process. Default is TRUE.}

\item{seed}{Seed for the random generator number. Parameter included for reproducibility purposes. Default is NULL (meaning no seed is assigned).}
}
\value{
A list containing the matrix of scores, the matrix of loadings, and a secondary list with extra information.
}
\description{
Performs cross validation on alpha parameter of fqpca
}
\examples{
# Generate fake dataset with 200 observations and 144 time points

x = matrix(rep(sin(seq(0, 2*pi, length.out=144)), 200), byrow=TRUE, nrow=200)
x = x + matrix(rnorm(200*144, 0, 0.4), nrow=200)

# Add missing observations
x[sample(200*144, as.integer(0.2*200*144))] = NA

cv_result = cross_validation_alpha(x, alpha_grid=c(0, 1e-15), n_folds=2)

}
