% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca-utils.R
\name{cross_validation_df}
\alias{cross_validation_df}
\title{CROSS VALIDATION OF DEGREES OF FREEDOM}
\usage{
cross_validation_df(
  data,
  colname = NULL,
  npc = 2,
  pve = NULL,
  quantile.value = 0.5,
  lambda.ridge = 0,
  n.folds = 3,
  return.models = TRUE,
  criteria = "points",
  periodic = TRUE,
  splines.df.grid = c(5, 10, 15, 20),
  tol = 0.001,
  n.iters = 20,
  method = "conquer",
  penalized = FALSE,
  weights = NULL,
  parallelized.scores = FALSE,
  num.cores = NULL,
  verbose.fqpca = FALSE,
  verbose.cv = TRUE,
  seed = NULL
)
}
\arguments{
\item{data}{An \eqn{(N \times T)} matrix, a tf object from the tidyfun package or a data.frame containing the functional data as a tf column.}

\item{colname}{The name of the column containing the functional data. Use only if data is a dataframe and colname is a column in the dataframe.}

\item{npc}{The number of estimated components.}

\item{pve}{Float between 0 and 1. Percentage of variability explained by components. This affects the number of components used in the curve reconstruction and error estimation. Set to NULL to avoid this behavior.}

\item{quantile.value}{The quantile considered.}

\item{lambda.ridge}{Hyper parameter controlling the penalization on the second derivative of the splines. It has effect only with \code{penalized=TRUE}. Experimantal component.}

\item{n.folds}{Number of folds to be used on cross validation.}

\item{return.models}{Should the list of all the models built be returned?}

\item{criteria}{Criteria used to divide the data. Valid values are \code{'rows'}, which considers the division based on full rows, or \code{'points'}, which considers the division based on points within the matrix.}

\item{periodic}{Boolean indicating if the data is expected to be periodic (start coincides with end) or not.}

\item{splines.df.grid}{Grid of possible values for the degrees of freedom.}

\item{tol}{Tolerance on the convergence of the algorithm.}

\item{n.iters}{Maximum number of iterations.}

\item{method}{Method used in the resolution of the quantile regression model. It currently accepts the methods \code{c('br', 'fn', 'pfn', 'sfn')} from \code{quantreg} package along with any available solver in \code{CVXR} package.}

\item{penalized}{Boolean indicating if the smoothness should be controlled using a second derivative penalty. This functionality is experimental and is much slower than the control of the smoothness using the degrees of freedom.}

\item{weights}{Optional array of weights used in penalized fqpca models. It must have dimension equal to (npc+1)}

\item{parallelized.scores}{Should the scores be computed in parallel? Experimantal component.}

\item{num.cores}{Number of cores to use in parallelized executions.}

\item{verbose.fqpca}{Boolean indicating verbosity of the fqpca function.}

\item{verbose.cv}{Boolean indicating verbosity of the cross-validation process.}

\item{seed}{Seed for the random generator number.}
}
\value{
A list containing the matrix of scores, the matrix of loadings, and a secondary list with extra information.
}
\description{
Performs cross validation on degrees of freedom parameter of fqpca
}
\examples{
# Generate fake dataset with 200 observations and 144 time points

Y <- matrix(rep(sin(seq(0, 2*pi, length.out = 144)), 200), byrow = TRUE, nrow = 200)
Y <- Y + matrix(rnorm(200*144, 0, 0.4), nrow = 200)

# Add missing observations
Y[sample(200*144, as.integer(0.2*200*144))] <- NA

cv_result <- cross_validation_df(data=Y, splines.df.grid = c(5, 10, 15), n.folds = 2)
}
