% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-factors.R
\name{explained_variance_ratio}
\alias{explained_variance_ratio}
\title{Explained variance ratio of the scores}
\usage{
explained_variance_ratio(scores)
}
\arguments{
\item{scores}{Matrix of scores.}
}
\value{
The percentage of variability each component is explaining.
}
\description{
Computes the percentage of explained variability based on the variance of the scores matrix
}
