% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosqr_fqpca.R
\name{fitted.fosqr_fqpca_object}
\alias{fitted.fosqr_fqpca_object}
\title{Fit Yhat}
\usage{
\method{fitted}{fosqr_fqpca_object}(object, pve = 0.95, ...)
}
\arguments{
\item{object}{An object output of the fosqr_fqpca function.}

\item{pve}{If smaller than 1, taken as percentage of explained variability used in Yhat estimation. If greater than 1, taken as number of components  used in Yhat estimation.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The matrix of fitted values
}
\description{
S3 method for class 'fosqr_fqpca_object'. Given a fosqr_fqpca_object model, estimates Yhat for different pve values.
}
\examples{

n.obs = 150
n.time = 144
time.grid <- seq(0, 2 * pi, length.out = n.time)

# Generate regressor and principal component functions
b1 = sin(time.grid)
pc1 = sin(2 * time.grid)

# Generate covariates and scores
regressors =  5 * matrix(rnorm(n.obs), ncol=1)
scores = matrix(10 * rnorm(n.obs), ncol = 1)
epsilon.matrix <- 0.1 * matrix(rnorm(n.obs * n.time), nrow = n.obs)

# Build dataset
Y = 100 + regressors[, 1] \%o\% b1 + scores[, 1] \%o\% pc1 + epsilon.matrix

# Add missing observations
Y[sample(n.obs*n.time, as.integer(0.2*n.obs*n.time))] <- NA

results <- fosqr_fqpca(Y = Y, regressors=regressors, npc = 1, quantile.value = 0.5, seed=1)
Yhat <- fitted(object = results, pve=0.95)
}
