% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosqr.R
\name{fitted.fosqr_object}
\alias{fitted.fosqr_object}
\title{Fit Yhat}
\usage{
\method{fitted}{fosqr_object}(object, ...)
}
\arguments{
\item{object}{An object output of the fosqr function.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The matrix of fitted values.
}
\description{
S3 method for class 'fosqr_object'. Given a fosqr_object model, extracts the fitted curves.
}
