% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosqr.R
\name{fosqr}
\alias{fosqr}
\title{FOSQR main function}
\usage{
fosqr(
  Y = NULL,
  regressors = NULL,
  quantile.value = 0.5,
  periodic = TRUE,
  splines.df = 10,
  splines.method = "quantreg",
  verbose = FALSE,
  seed = NULL
)
}
\arguments{
\item{Y}{An \eqn{(N \times T)} matrix or a tf object from the tidyfun package.}

\item{regressors}{An \eqn{(N \times P)} matrix of covariates.}

\item{quantile.value}{The quantile considered.}

\item{periodic}{Boolean indicating if the data is expected to be periodic (start coincides with end) or not.}

\item{splines.df}{Degrees of freedom for the splines.}

\item{splines.method}{Method used in the resolution of the splines quantile regression model. It currently accepts the methods \code{c('conquer', 'quantreg')}.}

\item{verbose}{Boolean indicating the verbosity.}

\item{seed}{Seed for the random generator number.}
}
\value{
fosqr_object
}
\description{
Solves the Function on Scalar Quantile Regression (FOSQR) model.
}
