% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosqr_fqpca.R
\name{fosqr_fqpca}
\alias{fosqr_fqpca}
\title{FOSQR - FQPCA main function}
\usage{
fosqr_fqpca(
  Y = NULL,
  regressors = NULL,
  npc = 2,
  quantile.value = 0.5,
  periodic = TRUE,
  splines.df = 10,
  splines.method = "conquer",
  tol = 0.001,
  max.iters = 20,
  verbose = FALSE,
  seed = NULL
)
}
\arguments{
\item{Y}{An \eqn{(N \times T)} matrix or a tf object from the tidyfun package.}

\item{regressors}{An \eqn{(N \times P)} matrix of covariates.}

\item{npc}{The number of estimated components on the FQPCA part (used for modeling residual correlation)}

\item{quantile.value}{The quantile considered.}

\item{periodic}{Boolean indicating if the data is expected to be periodic (start coincides with end) or not.}

\item{splines.df}{Degrees of freedom for the splines.}

\item{splines.method}{Method used in the resolution of the splines quantile regression model. It currently accepts the methods \code{c('conquer', 'quantreg')}.}

\item{tol}{Tolerance on the convergence of the algorithm.}

\item{max.iters}{Maximum number of iterations.}

\item{verbose}{Boolean indicating the verbosity.}

\item{seed}{Seed for the random generator number.}
}
\value{
fosqr_fqpca_object
}
\description{
Solves the Function on Scalar Quantile Regression (FOSQR) model
accounting for residual correlation based on Functional Principal Component
Analysis (FQPCA)
}
\examples{

n.obs = 150
n.time = 144
time.grid <- seq(0, 2 * pi, length.out = n.time)

# Generate regressor and principal component functions
b1 = sin(time.grid)
pc1 = sin(2 * time.grid)

# Generate covariates and scores
regressors =  5 * matrix(rnorm(n.obs), ncol=1)
scores = matrix(10 * rnorm(n.obs), ncol = 1)
epsilon.matrix <- 0.1 * matrix(rnorm(n.obs * n.time), nrow = n.obs)

# Build dataset
Y = 100 + regressors[, 1] \%o\% b1 + scores[, 1] \%o\% pc1 + epsilon.matrix

# Add missing observations
Y[sample(n.obs*n.time, as.integer(0.2*n.obs*n.time))] <- NA

results <- fosqr_fqpca(Y = Y, regressors = regressors, npc = 1, quantile.value = 0.5, seed=1)
}
