% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosqr_fqpca-utils.R
\name{fosqr_fqpca_cv_df}
\alias{fosqr_fqpca_cv_df}
\title{CROSS VALIDATION OF DEGREES OF FREEDOM}
\usage{
fosqr_fqpca_cv_df(
  Y = NULL,
  regressors = NULL,
  npc = 2,
  pve = 0.95,
  quantile.value = 0.5,
  periodic = TRUE,
  splines.df.grid = c(5, 10, 15),
  splines.method = "conquer",
  n.folds = 3,
  return.models = TRUE,
  criteria = "points",
  tol = 0.001,
  max.iters = 20,
  verbose.fosqr_fqpca = FALSE,
  verbose.cv = TRUE,
  seed = NULL
)
}
\arguments{
\item{Y}{An \eqn{(N \times T)} matrix or a tf object from the tidyfun package.}

\item{regressors}{An \eqn{(N \times P)} matrix of covariates.}

\item{npc}{The number of estimated components on the FQPCA part (used for modeling residual correlation)}

\item{pve}{Float between 0 and 1. Percentage of variability explained by components. This affects the number of components used in the curve reconstruction and error estimation. Set to NULL to avoid this behavior.}

\item{quantile.value}{The quantile considered.}

\item{periodic}{Boolean indicating if the data is expected to be periodic (start coincides with end) or not.}

\item{splines.df.grid}{Grid of possible values for the degrees of freedom.}

\item{splines.method}{Method used in the resolution of the splines quantile regression model. It currently accepts the methods \code{c('conquer', 'quantreg')}.}

\item{n.folds}{Number of folds to be used on cross validation.}

\item{return.models}{Should the list of all the models built be returned?}

\item{criteria}{Criteria used to divide the data. Valid values are \code{'rows'}, which considers the division based on full rows, or \code{'points'}, which considers the division based on points within the matrix.}

\item{tol}{Tolerance on the convergence of the algorithm.}

\item{max.iters}{Maximum number of iterations.}

\item{verbose.fosqr_fqpca}{Boolean indicating verbosity of the fqpca function.}

\item{verbose.cv}{Boolean indicating verbosity of the cross-validation process.}

\item{seed}{Seed for the random generator number.}
}
\value{
A list containing the matrices of scores, the matrices of loadings, and a secondary list with extra information.
}
\description{
Performs cross validation on degrees of freedom parameter of fosqr_fqpca
}
\examples{
n.obs = 150
n.time = 144
time.grid <- seq(0, 2 * pi, length.out = n.time)

# Generate regressor and principal component functions
b1 = sin(time.grid)
pc1 = sin(2 * time.grid)

# Generate covariates and scores
regressors =  5 * matrix(rnorm(n.obs), ncol=1)
scores = matrix(10 * rnorm(n.obs), ncol = 1)
epsilon.matrix <- 0.1 * matrix(rnorm(n.obs * n.time), nrow = n.obs)

# Build dataset
Y = 100 + regressors[, 1] \%o\% b1 + scores[, 1] \%o\% pc1 + epsilon.matrix

# Add missing observations
Y[sample(n.obs*n.time, as.integer(0.2*n.obs*n.time))] <- NA

cv_results <- fosqr_fqpca_cv_df(
    Y = Y, regressors = regressors,
    splines.df.grid = c(5, 10, 15), n.folds = 2)

}
