% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosqr_fqpca.R
\name{fosqrfqpca_compute_fosr_spline_coefficients}
\alias{fosqrfqpca_compute_fosr_spline_coefficients}
\title{Compute fosqr spline coefficients}
\usage{
fosqrfqpca_compute_fosr_spline_coefficients(
  Y.vector,
  Y.mask,
  regressors,
  spline.basis,
  quantile.value,
  method
)
}
\arguments{
\item{Y.vector}{vectorised version of Y, the \eqn{(N \times T)} matrix of observed time instants.}

\item{Y.mask}{Mask matrix of the same dimensions as Y indicating which observations in Y are known.}

\item{regressors}{Matrix of covariates.}

\item{spline.basis}{The spline basis matrix.}

\item{quantile.value}{The quantile considered.}

\item{method}{Method used in the resolution of the quantile regression model. It currently accepts the methods \code{c('conquer', 'quantreg')}.}
}
\value{
The matrix of spline coefficients.
}
\description{
Inner function to compute the spline coefficients of the fosqr-fqpca methodology
}
