% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosqr_fqpca.R
\name{fosqrfqpca_compute_loadings_fosqr}
\alias{fosqrfqpca_compute_loadings_fosqr}
\title{Compute loadings (aka principal components)}
\usage{
fosqrfqpca_compute_loadings_fosqr(spline.basis, spline.coefficients)
}
\arguments{
\item{spline.basis}{The spline basis matrix.}

\item{spline.coefficients}{the matrix of spline coefficients.}
}
\value{
The matrix of loadings
}
\description{
Inner function to compute the loadings of the fqpca methodology.
}
