% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosqr_fqpca.R
\name{fosqrfqpca_compute_spline_coefficients_variance_estimation}
\alias{fosqrfqpca_compute_spline_coefficients_variance_estimation}
\title{Compute spline coefficients}
\usage{
fosqrfqpca_compute_spline_coefficients_variance_estimation(
  Y.vector,
  Y.mask,
  scores,
  spline.basis,
  quantile.value,
  method
)
}
\arguments{
\item{Y.vector}{vectorised version of Y, the \eqn{(N \times T)} matrix of observed time instants.}

\item{Y.mask}{Mask matrix of the same dimensions as Y indicating which observations in Y are known.}

\item{scores}{Initial matrix of regressors and / or scores}

\item{spline.basis}{The spline basis matrix.}

\item{quantile.value}{The quantile considered.}

\item{method}{Method used in the resolution of the quantile regression model. It currently accepts the methods \code{c('conquer', 'quantreg')}.}
}
\value{
A list with the matrix of spline coefficients, the flat coefficient vector, and the tensor design matrix.
}
\description{
Inner function to compute the spline coefficients of the fosqr - fqpca methodology, additionally returning the tensor design matrix and flat coefficient vector needed to estimate the sandwich covariance via \code{get_kernel_cov}.
}
