% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosqr_fqpca.R
\name{fosqrfqpca_rotate_scores_and_loadings}
\alias{fosqrfqpca_rotate_scores_and_loadings}
\title{Rotation of fqpca loadings and scores}
\usage{
fosqrfqpca_rotate_scores_and_loadings(
  fqpca.loadings,
  fosqr.intercept,
  fqpca.scores
)
}
\arguments{
\item{fqpca.loadings}{Matrix of fqpca loadings.}

\item{fosqr.intercept}{population intercept}

\item{fqpca.scores}{Matrix of fqpca scores.}
}
\value{
The rotated matrices of loadings and scores and the rotation matrix.
}
\description{
Performs the rotation of loadings and scores in order to ensure the solution is unique.
}
