% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca_lasso.R
\name{fqpca_lasso}
\alias{fqpca_lasso}
\title{FQPCA}
\usage{
fqpca_lasso(
  x,
  n_components = 2,
  quantile_value = 0.5,
  periodic = TRUE,
  splines_df = 30,
  alpha_ridge = 1e-08,
  penalized_loadings = TRUE,
  tol = 0.001,
  n_iters = 30,
  solver = "SCS",
  alpha_lasso_scores = c(10^seq(-10, -1, length.out = 10), 0, 10^seq(1, 5, length.out =
    5)),
  verbose = FALSE,
  seed = 1
)
}
\arguments{
\item{x}{The N by T matrix of observed time instants}

\item{n_components}{The number of estimated components. Default is 2.}

\item{quantile_value}{The quantile considered. Default is the median 0.5.}

\item{periodic}{Boolean indicating if the data is expected to be periodic (start coincides with end) or not. Default is TRUE.}

\item{splines_df}{Degrees of freedom for the splines. It is recommended not to modify this parameter and control the smoothness via the hyper-parameter alpha_ridge. Default is 30.}

\item{alpha_ridge}{The hyperparameter controlling the penalization on the splines. This parameter has no effect if parameter penalized_loadings is set to FALSE. Default is 1e-16.}

\item{penalized_loadings}{Boolean indicating if the model should solve a penalized splines model or a non penalized splines model}

\item{tol}{Tolerance on the convergence of the algorithm. Smaller values can speed up computation but may affect the quality of the estimations. Default is 1e-3.}

\item{n_iters}{Maximum number of iterations. Default is 30.}

\item{solver}{Solver to be used by the CVXR package in the resolution of the penalized quantile regression models. The speed of the methodology can vary depending on the solver used. There are free alternatives available like 'SCS', 'ECOS' or 'OSQP'. It is also possible to use licensed programs like 'MOSEK' or 'GUROBI'. Default is 'SCS'.}

\item{alpha_lasso_scores}{Alpha hyper-parameter on lasso penalized quantile regression for the scores.}

\item{verbose}{Boolean indicating the verbosity of the function.}

\item{seed}{Seed for the random generator number. Parameter included for reproducibility purposes. Default is NULL (meaning no seed is assigned).}
}
\value{
A list containing the matrix of scores, the matrix of loadings, and a secondary list with extra information.
}
\description{
Solves the functional quantile principal component analysis methodology
}
\examples{
# Generate fake dataset with 150 observations and 144 time points

x = matrix(rep(sin(seq(0, 2*pi, length.out=144)), 150), byrow=TRUE, nrow=150)
x = x + matrix(rnorm(150*144, 0, 0.4), nrow=150)

# Add missing observations
x[sample(150*144, as.integer(0.2*150*144))] = NA

results = fqpca_lasso(x=x, n_components=2, quantile_value=0.5, alpha_ridge=1e-7)

loadings = results$loadings
scores = results$scores
}
