% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca_no_scaling.R
\name{fqpca_no_scaling}
\alias{fqpca_no_scaling}
\title{FQPCA}
\usage{
fqpca_no_scaling(
  x,
  n_components = 2,
  quantile_value = 0.5,
  lambda_ridge = 1e-12,
  periodic = TRUE,
  splines_df = 30,
  tol = 0.001,
  n_iters = 30,
  solver = "SCS",
  penalized = TRUE,
  method = "br",
  verbose = FALSE,
  seed = NULL
)
}
\arguments{
\item{x}{The N by T matrix of observed time instants}

\item{n_components}{The number of estimated components. Default is 2.}

\item{quantile_value}{The quantile considered. Default is the median 0.5.}

\item{lambda_ridge}{The hyperparameter controlling the penalization on the splines. This parameter has no effect if parameter penalized is set to FALSE. Default is 1e-12.}

\item{periodic}{Boolean indicating if the data is expected to be periodic (start coincides with end) or not. Default is TRUE.}

\item{splines_df}{Degrees of freedom for the splines. It is recommended not to modify this parameter and control the smoothness via the hyperparameter lambda_ridge. Default is 30.}

\item{tol}{Tolerance on the convergence of the algorithm. Smaller values can spped up computation but may affect the quality of the estimations. Default is 1e-3.}

\item{n_iters}{Maximum number of iterations. Default is 30.}

\item{solver}{Solver to be used by the CVXR package in the resolution of the penalized quantile regression models. The speed of the methodology can vary depending on the solver used. There are free alternatives available like 'SCS', 'ECOS' or 'OSQP'. It is also possible to use licensed programs like 'MOSEK' or 'GUROBI'. Default is 'SCS'.}

\item{penalized}{Boolean indicating if the model should solve a penalized splines model or a non penelized splines model}

\item{method}{Method to be used by the quantreg package when solving quantile regression models. This parameter has no effect if penalized is TRUE (default)}

\item{verbose}{Boolean indicating verbosity of the function. Default is FALSE.}

\item{seed}{Seed for the random generator number. Parameter included for reproducibility purposes. Default is NULL (meaning no seed is assigned).}
}
\value{
A list containing the matrix of scores, the matrix of loadings, and a secondary list with extra information.
}
\description{
Solves the functional quantile principal component analysis methodology
}
\examples{
# Generate fake dataset with 200 observations and 144 time points

x = matrix(rep(sin(seq(0, 2*pi, length.out=144)), 200), byrow=TRUE, nrow=200)
x = x + matrix(rnorm(200*144, 0, 0.4), nrow=200)

# Add missing observations
x[sample(200*144, as.integer(0.2*200*144))] = NA

results = fqpca_no_scaling(x=x, n_components=1, quantile_value=0.5, lambda_ridge=1e-12)

loadings = results$loadings
scores = results$scores
}
