% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca.R
\name{inner_loop}
\alias{inner_loop}
\title{Inner loop function of the fqpca methodology.}
\usage{
inner_loop(
  x,
  x_mask,
  Lambda0,
  spline_basis_i,
  quantile_value,
  lambda_ridge,
  R_block,
  solver,
  penalized,
  method
)
}
\arguments{
\item{x}{The N by T matrix of observed time instants.}

\item{x_mask}{Mask matrix of the same dimensions asx indicating wihch observations in x are known.}

\item{Lambda0}{Initial matrix of scores.}

\item{spline_basis_i}{A list containing, for each observation xi, the matrix of associated spline basis.}

\item{quantile_value}{The quantile considered. Default is the median 0.5.}

\item{lambda_ridge}{The hyperparameter controlling the penalization on the splines. This parameter has no effect if parameter penalized is set to FALSE. Default is 1e-12.}

\item{R_block}{Block diagonal matrix made of quadratic matrices used to apply a ridge based penalty.  This only has effect if parameter 'penalized' is TRUE.}

\item{solver}{Solver to be used by the CVXR package in the resolution of the penalized quantile regression models. The speed of the methodology can vary depending on the solver used. There are free alternatives available like 'SCS', 'ECOS' or 'OSQP'. It is also possible to use licensed programs like 'MOSEK' or 'GUROBI'. Default is 'SCS'. This parameter only has effect if parameter 'penalized' is TRUE.}

\item{penalized}{Boolean indicating if the model should solve a penalized splines model or a non penelized splines model}

\item{method}{Method to be used by the quantreg package when solving quantile regression models. This parameter has no effect if penalized is TRUE (default)}
}
\value{
The loadings and scores matrices, the matrix of spline coefficients and the objective function value at the current iteration.
}
\description{
Inner loop function of the fqpca methodology.
}
