% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca-utils.R
\name{is_rank_deficient}
\alias{is_rank_deficient}
\title{Check if a matrix is singular}
\usage{
is_rank_deficient(x)
}
\arguments{
\item{x}{Input matrix.}
}
\value{
Boolean (true if)
}
\description{
Given a matrix, return TRUE if singular
}
