% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca-utils.R
\name{kfold_cv_points}
\alias{kfold_cv_points}
\title{Split a given matrix Y into k-folds}
\usage{
kfold_cv_points(Y, folds = 3, seed = NULL)
}
\arguments{
\item{Y}{\eqn{(N \times P)} matrix of predictive variables.}

\item{folds}{Integer number indicating number of folds.}

\item{seed}{Seed for the random generator number.}
}
\value{
A list containing two inside lists, one for training and one for testing. The length of the inside lists is equal to the number of folds
}
\description{
Splits a given matrix Y into k-folds. This function keeps the dimensions of the original Y in both train and test and substitutes the values in both split by NAs
}
