% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfqpca.R
\name{mfqpca_check_input_data}
\alias{mfqpca_check_input_data}
\title{Check input data}
\usage{
mfqpca_check_input_data(data, group, colname)
}
\arguments{
\item{data}{An \eqn{(N \times T)} matrix, a tf object from the tidyfun package or a data.frame containing the functional data as a tf column.}

\item{group}{The grouping structure of the hierarchical data. If data is a dataframe this can be a name pointing to a column in the dataframe. If data is any other object it must be an array with length equal to the number of rows of data.}

\item{colname}{The name of the column containing the functional data. Use only if data is a dataframe and colname is a column in the dataframe.}
}
\value{
A list containing two elements:
1. An unnamed matrix of the functional data.
2. The processed group vector.
}
\description{
Check input data
}
