% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfqpca.R
\name{mfqpca_check_params}
\alias{mfqpca_check_params}
\title{Check input parameters}
\usage{
mfqpca_check_params(
  npc.between,
  npc.within,
  quantile.value,
  periodic,
  splines.df,
  splines.method,
  tol,
  max.iters,
  verbose,
  seed
)
}
\arguments{
\item{npc.between}{The number of estimated between components.}

\item{npc.within}{The number of estimated within components.}

\item{quantile.value}{The quantile considered.}

\item{periodic}{Boolean indicating if the data is expected to be periodic (start coincides with end) or not.}

\item{splines.df}{Degrees of freedom for the splines.}

\item{splines.method}{Method used in the resolution of the splines quantile regression model. It currently accepts the methods \code{c('conquer', 'quantreg')}.}

\item{tol}{Tolerance on the convergence of the algorithm.}

\item{max.iters}{Maximum number of iterations.}

\item{verbose}{Boolean indicating the verbosity.}

\item{seed}{Seed for the random generator number.}
}
\value{
No return
}
\description{
Check input parameters
}
