% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfqpca.R
\name{mfqpca_compute_loadings_between}
\alias{mfqpca_compute_loadings_between}
\title{Compute loadings (aka principal components)}
\usage{
mfqpca_compute_loadings_between(spline.basis, spline.coefficients)
}
\arguments{
\item{spline.basis}{The spline basis matrix.}

\item{spline.coefficients}{the matrix of spline coefficients.}
}
\value{
The matrix of loadings
}
\description{
Inner function to compute the loadings of the mfqpca methodology when using conquer or other intercept alternatives
}
