% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfqpca.R
\name{mfqpca_compute_scores_between}
\alias{mfqpca_compute_scores_between}
\title{Compute scores between}
\usage{
mfqpca_compute_scores_between(
  Y,
  Y.mask,
  group,
  intercept,
  loadings,
  quantile.value
)
}
\arguments{
\item{Y}{The \eqn{(N \times T)} matrix of observed time instants.}

\item{Y.mask}{Mask matrix of the same dimensions as Y indicating which observations in Y are known.}

\item{group}{An N dimensional array indicating the hierarchical structure of the data. Elements in the array with the same value indicate they are repeated measures of the same individual.}

\item{intercept}{A T dimensional vector of intercept values.}

\item{loadings}{Matrix of loading coefficients.}

\item{quantile.value}{The quantile considered.}
}
\value{
The matrix of between level scores.
}
\description{
Inner function to compute the between level scores of the mfqpca methodology.
}
