% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfqpca.R
\name{mfqpca_compute_scores_within}
\alias{mfqpca_compute_scores_within}
\title{Compute scores within}
\usage{
mfqpca_compute_scores_within(Y, Y.mask, loadings, quantile.value)
}
\arguments{
\item{Y}{The \eqn{(N \times T)} matrix of observed time instants.}

\item{Y.mask}{Mask matrix of the same dimensions as Y indicating which observations in Y are known.}

\item{loadings}{Matrix of loading coefficients.}

\item{quantile.value}{The quantile considered.}
}
\value{
The matrix of within level scores.
}
\description{
Inner function to compute the within level scores of the mfqpca methodology.
}
