% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfqpca.R
\name{mfqpca_compute_splines_between}
\alias{mfqpca_compute_splines_between}
\title{Compute spline coefficients between}
\usage{
mfqpca_compute_splines_between(
  Y.vector,
  Y.mask,
  scores,
  spline.basis,
  quantile.value,
  method
)
}
\arguments{
\item{Y.vector}{vectorised version of Y, the \eqn{(N \times T)} matrix of observed time instants.}

\item{Y.mask}{Mask matrix of the same dimensions as Y indicating which observations in Y are known.}

\item{scores}{Initial matrix of scores.}

\item{spline.basis}{The spline basis matrix.}

\item{quantile.value}{The quantile considered.}

\item{method}{Method used in the resolution of the quantile regression model. It currently accepts the methods \code{c('conquer', 'quantreg')}.}
}
\value{
The matrix of spline coefficients splines coefficients.
}
\description{
Inner function to compute the spline coefficients of the mfqpca methodology using quantreg
}
