% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfqpca.R
\name{mfqpca_reconstruct}
\alias{mfqpca_reconstruct}
\title{Helper function to compute fitted values}
\usage{
mfqpca_reconstruct(scores, loadings, n.comp)
}
\arguments{
\item{scores}{matrix of scores}

\item{loadings}{matrix of loadings}

\item{n.comp}{number of components used. If 0 then a matrix of 0s is returned.}
}
\description{
Helper function to compute fitted values
}
