% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfqpca.R
\name{mfqpca_rotate_scores_and_loadings_between}
\alias{mfqpca_rotate_scores_and_loadings_between}
\title{Rotation of fqpca loadings and scores}
\usage{
mfqpca_rotate_scores_and_loadings_between(scores, intercept, loadings)
}
\arguments{
\item{scores}{Matrix of scores.}

\item{intercept}{population intercept}

\item{loadings}{Matrix of loadings.}
}
\value{
The rotated matrices of loadings and scores and the rotation matrix.
}
\description{
Performs the rotation of loadings and scores in order to ensure the solution is unique.
}
