% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtain_npc}
\alias{obtain_npc}
\title{Obtain npc}
\usage{
obtain_npc(scores, pve)
}
\arguments{
\item{scores}{matrix of scores}

\item{pve}{If smaller than 1, taken as percentage of explained variability. If greater than 1, taken as number of components.}
}
\description{
Computes the required number of components to achieve the desired pve value
}
