% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pivot_long_x}
\alias{pivot_long_x}
\title{Pivot long a matrix}
\usage{
pivot_long_x(x)
}
\arguments{
\item{x}{Input matrix.}
}
\value{
The pivoted matrix.
}
\description{
Given a new matrix x of dimensions (n, m) pivot into a (n*m, 3) matrix indexing (row, column, value)
}
