% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca-utils.R
\name{pivot_wide_x}
\alias{pivot_wide_x}
\title{Pivot short a matrix}
\usage{
pivot_wide_x(x.long, dimensions = NULL)
}
\arguments{
\item{x.long}{Input matrix.}

\item{dimensions}{array of two elements indicating the dimensions of the original data.}
}
\value{
The pivoted matrix as a tibble
}
\description{
Given a (n*m, 3) matrix indexing (row, column, value) pivots it into an (n, m) matrix.
}
