% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosqr.R
\name{plot.fosqr_object}
\alias{plot.fosqr_object}
\title{Plot fosqr loading functions with ggplot2}
\usage{
\method{plot}{fosqr_object}(x, ...)
}
\arguments{
\item{x}{An object output of the fosqr function.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A ggplot object plotting the intercept and FOSQR loading curves.
}
\description{
S3 method for class 'fosqr_object'. Given a fosqr_object model, plots the fosqr loadings.
}
