% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca.R
\name{plot.fqpca}
\alias{plot.fqpca}
\title{Plot FQPCA loading functions}
\usage{
\method{plot}{fqpca}(x, ...)
}
\arguments{
\item{x}{An object output of the fqpca function.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The plot of loadings
}
\description{
\subsection{S3 method for class 'fqpca'}{

Given a fqpca object, plot the loading functions
}
}
\examples{
# Generate fake dataset with 200 observations and 144 time points

x = matrix(rep(sin(seq(0, 2*pi, length.out=144)), 200), byrow=TRUE, nrow=200)
x = x + matrix(rnorm(200*144, 0, 0.4), nrow=200)

# Add missing observations
x[sample(200*144, as.integer(0.2*200*144))] = NA

results = fqpca(x=x[1:150,], n_components=1, quantile_value=0.5)

plot(results)
}
