% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca.R
\name{plot.fqpca_object}
\alias{plot.fqpca_object}
\title{Plot FQPCA loading functions}
\usage{
\method{plot}{fqpca_object}(x, ...)
}
\arguments{
\item{x}{An object output of the fqpca function.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The plot of loadings
}
\description{
\subsection{S3 method for class 'fqpca'}{

Given a fqpca object, plot the loading functions
}
}
\examples{
# Generate fake dataset with 150 observations and 144 time points

x = matrix(rep(sin(seq(0, 2*pi, length.out=144)), 150), byrow=TRUE, nrow=150)
x = x + matrix(rnorm(150*144, 0, 0.4), nrow=150)

# Add missing observations
x[sample(150*144, as.integer(0.2*150*144))] = NA

results = fqpca(x=x, n_components=2, quantile_value=0.5)

plot(results)
}
