% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfa.R
\name{plot.qfa}
\alias{plot.qfa}
\title{Plot QFA loadings}
\usage{
\method{plot}{qfa}(x, ...)
}
\arguments{
\item{x}{An object output of the fqpca function.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The plot of loadings
}
\description{
\subsection{S3 method for class 'qfa'}{

Given a fqpca object, plot the loadings
}
}
\examples{
# Generate fake dataset with 200 observations and 144 time points

x = matrix(rep(sin(seq(0, 2*pi, length.out=144)), 200), byrow=TRUE, nrow=200)
x = x + matrix(rnorm(200*144, 0, 0.4), nrow=200)

results = qfa(x=x[1:150,], n_components=1, quantile_value=0.5)

plot(results)
}
