% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosqr_fqpca.R
\name{predict.fosqr_fqpca_object}
\alias{predict.fosqr_fqpca_object}
\title{Predict scores}
\usage{
\method{predict}{fosqr_fqpca_object}(object, newdata, newregressors, ...)
}
\arguments{
\item{object}{An object output of the fosqr_fqpca function.}

\item{newdata}{The N by T matrix of observed time instants to be tested, or the tf object.}

\item{newregressors}{The N by p matrix of regressors to be tested}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The predicted matrix of scores.
}
\description{
S3 method for class 'fosqr_fqpca_object' Given a new matrix Y, predicts the value of the scores associated to the given matrix.
}
\examples{

n.obs = 150
n.time = 144
time.grid <- seq(0, 2 * pi, length.out = n.time)

# Generate regressor and principal component functions
b1 = sin(time.grid)
pc1 = sin(2 * time.grid)

# Generate covariates and scores
regressors =  5 * matrix(rnorm(n.obs), ncol=1)
scores = matrix(10 * rnorm(n.obs), ncol = 1)
epsilon.matrix <- 0.1 * matrix(rnorm(n.obs * n.time), nrow = n.obs)

# Build dataset
Y = 100 + regressors[, 1] \%o\% b1 + scores[, 1] \%o\% pc1 + epsilon.matrix

# Add missing observations
Y[sample(n.obs*n.time, as.integer(0.2*n.obs*n.time))] <- NA
results <- fosqr_fqpca(
    Y = Y[1:100, ],
    regressors = regressors[1:100, , drop=FALSE],
    npc = 1, quantile.value = 0.5, seed=1)

predictions <- predict(object = results, newdata = Y[101:150,],
    newregressors = regressors[101:150, , drop=FALSE])

}
