% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosqr.R
\name{predict.fosqr_object}
\alias{predict.fosqr_object}
\title{Predict FOSQR values}
\usage{
\method{predict}{fosqr_object}(object, newregressors, ...)
}
\arguments{
\item{object}{An object output of the fosqr function.}

\item{newregressors}{The N by p matrix of regressors to be tested.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The predicted matrix of fitted curves.
}
\description{
S3 method for class 'fosqr_object'. Predicts the fitted curves associated to new regressors.
}
