% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfa.R
\name{predict.qfa}
\alias{predict.qfa}
\title{Predict QFA}
\usage{
\method{predict}{qfa}(object, newdata, ...)
}
\arguments{
\item{object}{An object output of the qfa function.}

\item{newdata}{The N by T matrix of observed time instants to be tested.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The normalized matrix of scores.
}
\description{
\subsection{S3 method for class 'qfa'}{

Given a new matrix x, predicts the value of the scores associated to the given matrix
}
}
\examples{
# Generate fake dataset with 200 observations and 144 time points

x = matrix(rep(sin(seq(0, 2*pi, length.out=144)), 200), byrow=TRUE, nrow=200)
x = x + matrix(rnorm(200*144, 0, 0.4), nrow=200)

results = qfa(x=x[1:150,], n_components=1, quantile_value=0.5)

predictions = predict(results, newdata=x[151:200,])
}
