% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca.R
\name{predict_scores_fqpca}
\alias{predict_scores_fqpca}
\title{Predict FQPCA}
\usage{
predict_scores_fqpca(fqpca_object, x)
}
\arguments{
\item{fqpca_object}{An object output of the fqpca function.}

\item{x}{The N by T matrix of observed time instants to be tested.}
}
\value{
The normalized matrix of scores.
}
\description{
Given a new matrix x, predicts the value of the scores associated to the given matrix
}
\examples{
# Generate fake dataset with 200 observations and 144 time points

x = matrix(rep(sin(seq(0, 2*pi, length.out=144)), 200), byrow=TRUE, nrow=200)
x = x + matrix(rnorm(200*144, 0, 0.4), nrow=200)

# Add missing observations
x[sample(200*144, as.integer(0.2*200*144))] = NA

results = fqpca(x=x[1:150,], n.components=1, quantile.value=0.5)

predictions = predict_scores_fqpca(results, x=x[151:200,])
}
