% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfa.R
\name{predict_scores_qfa}
\alias{predict_scores_qfa}
\title{Predict QFA}
\usage{
predict_scores_qfa(qfa_object, x)
}
\arguments{
\item{qfa_object}{An object output of the qfa function.}

\item{x}{The N by T matrix of observed time instants to be tested.}
}
\value{
The normalized matrix of scores.
}
\description{
Predict QFA
}
\examples{
# Generate fake dataset with 200 observations and 144 time points

x = matrix(rep(sin(seq(0, 2*pi, length.out=144)), 200), byrow=TRUE, nrow=200)
x = x + matrix(rnorm(200*144, 0, 0.4), nrow=200)

results = qfa(x=x[1:150,], n.components=1, quantile.value=0.5)

predictions = predict_scores_qfa(results, x=x[151:200,])
}
