% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proportion_under_quantile}
\alias{proportion_under_quantile}
\title{Proportion of points under quantile}
\usage{
proportion_under_quantile(Y, Y.pred)
}
\arguments{
\item{Y}{The N by T matrix of observed time instants}

\item{Y.pred}{N by T matrix of predicted time instants}
}
\value{
the proportion of points under the estimated quantile function
}
\description{
Proportion of points under quantile
}
