% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfa.R
\name{qfa}
\alias{qfa}
\title{QFA}
\usage{
qfa(
  x,
  n_components,
  quantile_value = 0.5,
  tol = 0.001,
  n_iters = 100,
  verbose = TRUE,
  seed = NULL
)
}
\arguments{
\item{x}{The N by T matrix of observed time instants}

\item{n_components}{The number of estimated components. Default is 2.}

\item{quantile_value}{The quantile considered. Default is the median 0.5.}

\item{tol}{Tolerance on the convergence of the algorithm. Smaller values can spped up computation but may affect the quality of the estimations. Default is 1e-3.}

\item{n_iters}{Maximum number of iterations. Default is 30.}

\item{verbose}{Boolean indicating verbosity of the function. Default is FALSE.}

\item{seed}{Seed for the random generator number. Parameter included for reproducibility purposes. Default is NULL (meaning no seed is assigned).}
}
\value{
A list containing the matrix of scores, the matrix of loadings, and a secondary list with extra information.
}
\description{
Solves the Quantile factor analysis methodology.
}
\examples{
# Generate fake dataset with 200 observations and 144 time points

x = matrix(rep(sin(seq(0, 2*pi, length.out=144)), 200), byrow=TRUE, nrow=200)
x = x + matrix(rnorm(200*144, 0, 0.4), nrow=200)

results = qfa(x=x, n_components=1, quantile_value=0.5)

loadings = results$loadings
scores = results$scores
}
