% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qr}
\alias{qr}
\title{Quantile regression model}
\usage{
qr(x, y, quantile_value = 0.5, solver = "SCS")
}
\arguments{
\item{x}{N times p matrix of predictive variables.}

\item{y}{N times 1 vector of response.}

\item{quantile_value}{The quantile considered. Default is the median 0.5.}

\item{solver}{Solver to be used by the CVXR package in the resolution of the penalized quantile regression models. The speed of the methodology can vary depending on the solver used. There are free alternatives available like 'SCS', 'ECOS' or 'OSQP'. It is also possible to use licensed programs like 'MOSEK' or 'GUROBI'. Default is 'SCS'.}
}
\value{
Beta coefficients of the penalized quantile regression model.
}
\description{
Uses the CVXR library to solve a quantile regression model
}
