% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca.R
\name{qr_ridge}
\alias{qr_ridge}
\title{Penalized quantile regression}
\usage{
qr_ridge(
  x,
  y,
  R,
  quantile.value = 0.5,
  lambda = 0.001,
  verbose = FALSE,
  solver = "SCS"
)
}
\arguments{
\item{x}{N times p matrix of predictive variables.}

\item{y}{N times 1 vector of response.}

\item{R}{Quadratic matrix used to apply a ridge based penalty.}

\item{quantile.value}{The quantile considered. Default is the median 0.5.}

\item{lambda}{The hyperparameter controlling the penalization. Default is 1e-3.}

\item{verbose}{Boolean indicating verbosity of the function. Default is FALSE.}

\item{solver}{Solver to be used by the CVXR package in the resolution of the penalized quantile regression models. The speed of the methodology can vary depending on the solver used. There are free alternatives available like 'SCS', 'ECOS' or 'OSQP'. It is also possible to use licensed programs like 'MOSEK' or 'GUROBI'. Default is 'SCS'.}
}
\value{
Beta coefficients of the penalized quantile regression model.
}
\description{
Uses the CVXR library to solve a penalized quantile regression model
}
