% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quantile_error}
\alias{quantile_error}
\title{Quantile error computation}
\usage{
quantile_error(x, x_pred, quantile_value)
}
\arguments{
\item{x}{The N by T matrix of observed time instants}

\item{x_pred}{N by T matrix of predicted time instants}

\item{quantile_value}{The quantile considered.}
}
\value{
the quantile error between the two matrices
}
\description{
Quantile error computation
}
