% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quantile_function}
\alias{quantile_function}
\title{QR objective value}
\usage{
quantile_function(x, quantile_value = 0.5)
}
\arguments{
\item{x}{The vector.}

\item{quantile_value}{The quantile considered. Default is the median 0.5.}
}
\value{
Objective value function for a quantile regression model.
}
\description{
Objective value function of a quantile regression model
}
