% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca-utils.R
\name{quantile_regression}
\alias{quantile_regression}
\title{Quantile regression model}
\usage{
quantile_regression(x, y, quantile.value = 0.5, solver = "SCS")
}
\arguments{
\item{x}{\eqn{(N \times P)} matrix of predictive variables.}

\item{y}{\eqn{(N \times 1)} vector of response.}

\item{quantile.value}{The quantile considered.}

\item{solver}{Solver to be used by the \code{CVXR} package}
}
\value{
Beta coefficients of the quantile regression model.
}
\description{
Uses the CVXR library to solve a quantile regression model
}
