% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quantile_regression_ridge}
\alias{quantile_regression_ridge}
\title{Penalized quantile regression}
\usage{
quantile_regression_ridge(
  x,
  y,
  R,
  quantile.value = 0.5,
  lambda = 0.001,
  solver = "SCS"
)
}
\arguments{
\item{x}{N times p matrix of predictive variables.}

\item{y}{N times 1 vector of response.}

\item{R}{Quadratic matrix used to apply a ridge based penalty.}

\item{quantile.value}{Default<-0.5. The quantile considered.}

\item{lambda}{The hyperparameter controlling the penalization. Default is 1e-3.}

\item{solver}{Default<-'SCS'. Solver to be used by the CVXR package}
}
\value{
Beta coefficients of the penalized quantile regression model.
}
\description{
Uses the CVXR library to solve a penalized quantile regression model
}
