% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-factors.R
\name{reconstruct_scores_loadings}
\alias{reconstruct_scores_loadings}
\title{Truncated reconstruction from scores and loadings}
\usage{
reconstruct_scores_loadings(scores, loadings, n.comp)
}
\arguments{
\item{scores}{matrix of scores}

\item{loadings}{matrix of loadings}

\item{n.comp}{number of components used. If 0 then a matrix of 0s is returned.}
}
\description{
Helper function to compute fitted values from the first \code{n.comp} components.
}
