% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca.R
\name{rotate_scores_and_loadings}
\alias{rotate_scores_and_loadings}
\title{Rotation of fqpca loadings and scores}
\usage{
rotate_scores_and_loadings(loadings, scores)
}
\arguments{
\item{loadings}{Matrix of loadings.}

\item{scores}{Matrix of scores.}
}
\value{
The rotated matrices of loadings and scores and the rotation matrix.
}
\description{
Performs the rotation of loadings and scores in order to ensure the solution is unique.
}
