% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-factors.R
\name{select_npc}
\alias{select_npc}
\title{Select the number of components}
\usage{
select_npc(scores, pve, npc = NULL)
}
\arguments{
\item{scores}{matrix of scores}

\item{pve}{Percentage of explained variability, a number between 0 and 1. Set to NULL to use all components.}

\item{npc}{Number of components. If provided (non NULL), it supersedes \code{pve}. Must be an integer between 0 and the number of components in \code{scores}.}
}
\description{
Computes the required number of components to achieve the desired pve value
}
