% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosqr_fqpca.R
\name{check_regressors}
\alias{check_regressors}
\title{Checks input data}
\usage{
check_regressors(regressors)
}
\arguments{
\item{regressors}{An \eqn{(N \times P)} matrix}
}
\value{
The prepared regressors matrix
}
\description{
Checks that regressors is valid and formats it.
}
