% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca.R
\name{compute_scores}
\alias{compute_scores}
\title{Compute scores}
\usage{
compute_scores(x, x_mask, loadings, quantile_value, method)
}
\arguments{
\item{x}{The N by T matrix of observed time instants.}

\item{x_mask}{Mask matrix of the same dimensions as x indicating wich observations in x are known.}

\item{loadings}{Matrix of loading coefficients}

\item{quantile_value}{The quantile considered.}

\item{method}{Method to be used by the quantreg package when solving quantile regression models. This parameter has no effect if penalized is TRUE (default)}
}
\value{
The matrix of scores
}
\description{
Inner function to compute the scores of the fqpca methodology.
}
