% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca.R
\name{compute_scores}
\alias{compute_scores}
\title{Compute scores}
\usage{
compute_scores(Y, Y_mask, loadings, quantile.value)
}
\arguments{
\item{Y}{The N by T matrix of observed time instants.}

\item{Y_mask}{Mask matrix of the same dimensions as Y indicating wich observations in Y are known.}

\item{loadings}{Matrix of loading coefficients}

\item{quantile.value}{The quantile considered.}
}
\value{
The matrix of scores
}
\description{
Inner function to compute the scores of the fqpca methodology.
}
