% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca.R
\name{compute_spline_coefficients}
\alias{compute_spline_coefficients}
\title{Compute splines coefficients}
\usage{
compute_spline_coefficients(
  x,
  x_mask,
  scores,
  spline_basis,
  quantile_value,
  alpha_ridge,
  R_block,
  solver,
  penalized_loadings
)
}
\arguments{
\item{x}{The N by T matrix of observed time instants.}

\item{x_mask}{Mask matrix of the same dimensions as x indicating wich observations in x are known.}

\item{scores}{Initial matrix of scores.}

\item{spline_basis}{The spline basis matrix.}

\item{quantile_value}{The quantile considered. Default is the median 0.5.}

\item{alpha_ridge}{The hyper-parameter controlling the penalization on the splines. This parameter has no effect if parameter penalized is set to FALSE. Default is 1e-12.}

\item{R_block}{Block diagonal matrix made of quadratic matrices used to apply a ridge based penalty.  This only has effect if parameter 'penalized' is TRUE.}

\item{solver}{Solver to be used by the CVXR package in the resolution of the penalized quantile regression models. The speed of the methodology can vary depending on the solver used. There are free alternatives available like 'SCS', 'ECOS' or 'OSQP'. It is also possible to use licensed programs like 'MOSEK' or 'GUROBI'. Default is 'SCS'. This parameter only has effect if parameter 'penalized' is TRUE.}

\item{penalized_loadings}{Boolean indicating if the model should solve a penalized splines model or a non penalized splines model}
}
\value{
The matrix of spline coefficients splines coefficients
}
\description{
Inner function to compute the splines coefficients of the fqpca methodology.
}
