% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca.R
\name{compute_spline_coefficients}
\alias{compute_spline_coefficients}
\title{Compute spline coefficients}
\usage{
compute_spline_coefficients(
  penalized,
  Y,
  Y.mask,
  scores,
  spline.basis,
  quantile.value,
  method,
  lambda.ridge
)
}
\arguments{
\item{penalized}{Boolean indicating if the smoothness should be controlled using a second derivative penalty.}

\item{Y}{\eqn{(N \times T)} matrix of observed time instants.}

\item{Y.mask}{Mask matrix of the same dimensions as Y indicating wich observations in Y are known.}

\item{scores}{Initial matrix of scores.}

\item{spline.basis}{The spline basis matrix.}

\item{quantile.value}{The quantile considered.}

\item{method}{Method used in the resolution of the quantile regression model. It currently accepts the methods \code{c('conquer', 'quantreg')}.}

\item{lambda.ridge}{Hyper parameter controlling the penalization on the second derivative of the splines. It has effect only with \code{penalized=TRUE} and \code{method='conquer'}.}
}
\value{
The matrix of spline coefficients splines coefficients.
}
\description{
Inner function to compute the spline coefficients of the fqpca methodology.
}
