% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosqr_fqpca.R
\name{compute_var_sandwich}
\alias{compute_var_sandwich}
\title{Compute sandwich estimator variance associated to fosqr loadings}
\usage{
compute_var_sandwich(cov.model, n.regressors, spline.basis)
}
\arguments{
\item{cov.model}{fosqr-fqpca splines quantile regression model}

\item{n.regressors}{number of covariates}

\item{spline.basis}{spline basis matrix}
}
\value{
A matrix of variances associated to the fosqr loadings
}
\description{
Given a fosqr-fqpca splines coefficients model, estimates variances associated to fosqr loadings.
}
