% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca-utils.R
\name{create_folds}
\alias{create_folds}
\title{Split a given matrix Y into k-folds}
\usage{
create_folds(Y, criteria = "points", folds = 3, seed = NULL)
}
\arguments{
\item{Y}{\eqn{(N \times P)} matrix of predictive variables.}

\item{criteria}{Criteria used to divide the data. Valid values are \code{'rows'}, which considers the division based on full rows, or \code{'points'}, which considers the division based on points within the matrix.}

\item{folds}{Integer number indicating number of folds}

\item{seed}{Seed for the random generator number.}
}
\value{
A list containing two inside lists, one for training and one for testing. The length of the inside lists is equal to the number of folds
}
\description{
Splits a given matrix Y into k-folds. based on two possible criteria: either based on the total number of rows or the total number of data points.
}
\examples{
# Generate a small matrix
Y <- matrix(rnorm(50), nrow = 10)

# Divide based on full rows
kfolds_rows <- create_folds(Y, criteria = 'rows', folds = 3, seed = 1)
Y.train.list <- kfolds_rows$Y.train.list
Y.test.list <- kfolds_rows$Y.test.list

kfolds_points <- create_folds(Y, criteria<-'points', folds<-3, seed<-1)
Y.train.list <- kfolds_rows$Y.train.list
Y.test.list <- kfolds_rows$Y.test.list

}
