% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cross_validation_alpha}
\alias{cross_validation_alpha}
\title{CROSS VALIDATION}
\usage{
cross_validation_alpha(
  x,
  n_components = 2,
  quantile_value = 0.5,
  alpha_grid = c(0, 1e-16, 1e-14),
  n_folds = 3,
  criteria = "points",
  periodic = TRUE,
  splines_df = 10,
  tol = 0.001,
  n_iters = 50,
  method = "fn",
  verbose_fqpca = FALSE,
  verbose_cv = TRUE,
  seed = NULL
)
}
\arguments{
\item{x}{The N by T matrix of observed time instants}

\item{n_components}{Default=2. The number of estimated components.}

\item{quantile_value}{Default=0.5. The quantile considered.}

\item{alpha_grid}{Default=c(0, 1e-16, 1e-14). An array containing the list of
possible alpha values  (these should be always positive
numbers).}

\item{n_folds}{Default=3. Number of folds to be used on cross validation.}

\item{criteria}{Default='points'. Criteria used to divide the data.
Valid values are 'rows', which considers the division based
on full rows, or 'points', which considers the division based
on points within the matrix. Default is points}

\item{periodic}{Default=TRUE. Boolean indicating if the data is expected to
be periodic (start coincides with end) or not.}

\item{splines_df}{Default=10. Degrees of freedom for the splines.}

\item{tol}{Default=1e-3. Tolerance on the convergence of the algorithm.
Smaller values can speed up computation but may affect the
quality of the estimations.}

\item{n_iters}{Default=30. Maximum number of iterations.}

\item{method}{Default='fn'. Method used in the resolution of the quantile
regression model. It currently accepts the methods
c('br', 'fn', 'pfn', 'sfn') along with any available solver in
CVXR package like the free 'SCS' or the comercial 'MOSEK'.}

\item{verbose_fqpca}{Default=FALSE. Boolean indicating verbosity of the
fqpca function.}

\item{verbose_cv}{Default=TRUE. Boolean indicating verbosity of the cross
validation process.}

\item{seed}{Default=NULL. Seed for the random generator number.
Parameter included for reproducibility purposes.}
}
\value{
A list containing the matrix of scores, the matrix of loadings,
and a secondary list with extra information.
}
\description{
Performs cross validation on alpha parameter of fqpca
}
\examples{
# Generate fake dataset with 200 observations and 144 time points

x = matrix(rep(sin(seq(0, 2*pi, length.out=144)), 200), byrow=TRUE, nrow=200)
x = x + matrix(rnorm(200*144, 0, 0.4), nrow=200)

# Add missing observations
x[sample(200*144, as.integer(0.2*200*144))] = NA

cv_result = cross_validation_alpha(x, alpha_grid=c(0, 1e-15), n_folds=2)

}
