% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cross_validation_df}
\alias{cross_validation_df}
\title{CROSS VALIDATION OF DEGREES OF FREEDOM}
\usage{
cross_validation_df(
  Y,
  npc = 2,
  quantile.value = 0.5,
  alpha.ridge = 0,
  n.folds = 3,
  criteria = "points",
  periodic = TRUE,
  splines.df.grid = c(5, 10, 15, 20),
  tol = 0.001,
  n.iters = 50,
  method = "fn",
  verbose.fqpca = FALSE,
  verbose.cv = TRUE,
  seed = NULL
)
}
\arguments{
\item{Y}{The object storing the functional data. This can be either a matrix
data.frame or tibble of dimensions N by T, or a tf object from the
tidyfun package}

\item{npc}{Default=2. The number of estimated components.}

\item{quantile.value}{Default=0.5. The quantile considered.}

\item{alpha.ridge}{Default=0. Hyper parameter controlling the penalization
on the second derivative of the splines. It has effect
only with CVXR solvers. Large values are associated with
smoother results. This component is experimental and may
lead to computational issues.}

\item{n.folds}{Default=3. Number of folds to be used on cross validation.}

\item{criteria}{Default='points'. Criteria used to divide the data.
Valid values are 'rows', which considers the division based
on full rows, or 'points', which considers the division based
on points within the matrix. Default is points}

\item{periodic}{Default=TRUE. Boolean indicating if the data is expected to
be periodic (start coincides with end) or not.}

\item{splines.df.grid}{Default=c(5, 10, 15, 20). Grid of degrees of freedom
for the splines.}

\item{tol}{Default=1e-3. Tolerance on the convergence of the algorithm.
Smaller values can speed up computation but may affect the
quality of the estimations.}

\item{n.iters}{Default=30. Maximum number of iterations.}

\item{method}{Default='fn'. Method used in the resolution of the quantile
regression model. It currently accepts the methods
c('br', 'fn', 'pfn', 'sfn') from quantreg package along with
any available solver in CVXR package like the free 'SCS'
or the commercial 'MOSEK'.}

\item{verbose.fqpca}{Default=FALSE. Boolean indicating verbosity of the
fqpca function.}

\item{verbose.cv}{Default=TRUE. Boolean indicating verbosity of the cross
validation process.}

\item{seed}{Default=NULL. Seed for the random generator number.
Parameter included for reproducibility purposes.}
}
\value{
A list containing the matrix of scores, the matrix of loadings,
and a secondary list with extra information.
}
\description{
Performs cross validation on degrees of freedom parameter of fqpca
}
\examples{
# Generate fake dataset with 200 observations and 144 time points

Y = matrix(rep(sin(seq(0, 2*pi, length.out=144)), 200), byrow=TRUE, nrow=200)
Y = Y + matrix(rnorm(200*144, 0, 0.4), nrow=200)

# Add missing observations
Y[sample(200*144, as.integer(0.2*200*144))] = NA

cv_result = cross_validation_df(Y, splines.df.grid=c(5, 10, 15), n.folds=2)

}
