% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosqr_fqpca.R
\name{fosqrfqpca_build_splines_coefficients_matrix}
\alias{fosqrfqpca_build_splines_coefficients_matrix}
\title{Build spline coefficients matrix}
\usage{
fosqrfqpca_build_splines_coefficients_matrix(model, npc)
}
\arguments{
\item{model}{The output from fosqrfqpca_compute_spline_coefficients}

\item{npc}{number of pcs}
}
\value{
The matrix of spline coefficients
}
\description{
Inner function to build the spline coefficients matrix.
}
