% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosqr_fqpca.R
\name{fosqrfqpca_compute_spline_coefficients_variance_estimation}
\alias{fosqrfqpca_compute_spline_coefficients_variance_estimation}
\title{Compute spline coefficients}
\usage{
fosqrfqpca_compute_spline_coefficients_variance_estimation(
  Y.vector,
  Y.mask,
  scores,
  spline.basis,
  quantile.value
)
}
\arguments{
\item{Y.vector}{vectorised version of Y, the \eqn{(N \times T)} matrix of observed time instants.}

\item{Y.mask}{Mask matrix of the same dimensions as Y indicating which observations in Y are known.}

\item{scores}{Initial matrix of regressors and / or scores}

\item{spline.basis}{The spline basis matrix.}

\item{quantile.value}{The quantile considered.}
}
\value{
The matrix of spline coefficients.
}
\description{
Inner function to compute the spline coefficients of the fosqr - fqpca methodology using quantreg to estimate the variance for the sandwich estimator.
}
