% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kfold_cv_rows}
\alias{kfold_cv_rows}
\title{Split a given matrix X into k-folds}
\usage{
kfold_cv_rows(x, folds = 3, seed = NULL)
}
\arguments{
\item{x}{N times p matrix of predictive variables.}

\item{folds}{Integer number indicating number of folds}

\item{seed}{Seed for the random generator number. Parameter included for reproducibility purposes. Default is NULL (meaning no seed is assigned).}
}
\value{
A list containing two inside lists, one for training and one for testing. The length of the inside lists is equal to the number of folds
}
\description{
Splits full rows of a given matrix X into k-folds
}
