% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca.R
\name{plot.fqpca_object}
\alias{plot.fqpca_object}
\title{Plot fqpca loading functions with ggplot2}
\usage{
\method{plot}{fqpca_object}(x, pve = 0.99, ...)
}
\arguments{
\item{x}{An object output of the fqpca function.}

\item{pve}{Percentage of explained variability to determine the number of components.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A ggplot object plotting the intercept and FQPC curves.
}
\description{
S3 method for class 'fqpca_object'. Given a fqpca object, plot
the loading functions using ggplot2 facets.
}
\examples{

n.obs = 150
n.time = 144

# Generate scores
c1.vals = rnorm(n.obs)
c2.vals = rnorm(n.obs)

# Generate pc's
pc1 = sin(seq(0, 2*pi, length.out = n.time))
pc2 = cos(seq(0, 2*pi, length.out = n.time))

# Generate data
Y <- c1.vals * matrix(pc1, nrow = n.obs, ncol=n.time, byrow = TRUE) +
c2.vals * matrix(pc2, nrow = n.obs, ncol=n.time, byrow = TRUE) +
matrix(seq(1, 10, length.out=n.time), nrow = n.obs, ncol=n.time, byrow = TRUE)

# Add noise
Y <- Y + matrix(rnorm(n.obs * n.time, 0, 0.4), nrow = n.obs)

# Add missing observations
Y[sample(n.obs*n.time, as.integer(0.2*n.obs*n.time))] <- NA

results <- fqpca(data = Y, npc = 2, quantile.value = 0.5, periodic = FALSE, seed=1)

plot(results)
}
