% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca.R
\name{predict.fqpca_object}
\alias{predict.fqpca_object}
\title{Predict FQPCA}
\usage{
\method{predict}{fqpca_object}(object, newdata, ...)
}
\arguments{
\item{object}{An object output of the fqpca function.}

\item{newdata}{The N by T matrix of observed time instants to be tested.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The normalized matrix of scores.
}
\description{
\subsection{S3 method for class 'fqpca_object'}{

Given a new matrix x, predicts the value of the scores associated to the given matrix
}
}
\examples{
# Generate fake dataset with 150 observations and 144 time points

x = matrix(rep(sin(seq(0, 2*pi, length.out=144)), 150), byrow=TRUE, nrow=150)
x = x + matrix(rnorm(150*144, 0, 0.4), nrow=150)

# Add missing observations
x[sample(150*144, as.integer(0.2*150*144))] = NA

results = fqpca(x=x[1:100,], n_components=2, quantile_value=0.5)

predictions = predict(object=results, newdata=x[101:150,])
}
