% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca-utils.R
\name{proportion_under_quantile}
\alias{proportion_under_quantile}
\title{Proportion of points under quantile}
\usage{
proportion_under_quantile(Y, Y.pred)
}
\arguments{
\item{Y}{\eqn{(N \times T)} matrix of observed time instants.}

\item{Y.pred}{\eqn{(N \times T)} matrix of predicted time instants.}
}
\value{
The proportion of points under the estimated quantile function
}
\description{
Proportion of points under quantile
}
