% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqpca-utils.R
\name{quantile_error}
\alias{quantile_error}
\title{Quantile error computation}
\usage{
quantile_error(Y, Y.pred, quantile.value)
}
\arguments{
\item{Y}{\eqn{(N \times T)} matrix of observed time instants.}

\item{Y.pred}{\eqn{(N \times T)} matrix of predicted time instants.}

\item{quantile.value}{The quantile considered.}
}
\value{
The quantile error between the two matrices
}
\description{
Quantile error computation
}
