% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quantile_function}
\alias{quantile_function}
\title{Quantile regression objective value}
\usage{
quantile_function(x, quantile.value = 0.5)
}
\arguments{
\item{x}{The vector.}

\item{quantile.value}{Default<-0.5. The quantile considered.}
}
\value{
Objective value function for a quantile regression model.
}
\description{
Quantile regression objective value
}
