% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quantile_regression}
\alias{quantile_regression}
\title{Quantile regression model}
\usage{
quantile_regression(x, y, quantile.value = 0.5, solver = "SCS")
}
\arguments{
\item{x}{N times p matrix of predictive variables.}

\item{y}{N times 1 vector of response.}

\item{quantile.value}{Default<-0.5. The quantile considered.}

\item{solver}{Default<-'SCS'. Solver to be used by the CVXR package}
}
\value{
Beta coefficients of the penalized quantile regression model.
}
\description{
Uses the CVXR library to solve a quantile regression model
}
