% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{train_test_split}
\alias{train_test_split}
\title{Split a given matrix X into train / test}
\usage{
train_test_split(
  Y,
  criteria = "points",
  train.pct = NULL,
  train.size = NULL,
  seed = NULL
)
}
\arguments{
\item{Y}{N times p matrix of predictive variables.}

\item{criteria}{Default<-'points'. Criteria used to divide the data. Valid values are 'rows', which considers the division based on full rows, or 'points', which considers the division based on points within the matrix.}

\item{train.pct}{Default<-NULL. Float number indicating the \% of rows used for training. This takes precedence over \code{train.size}.}

\item{train.size}{Default<-NULL. Integer number indicating number of rows used for training. train.size is superseded by train.pct}

\item{seed}{Default<-NULL. Seed for the random generator number.}
}
\value{
a list containing a matrix x_train and a matrix x_test
}
\description{
Splits a given matrix into a train / test split based on two posible
criterias: either based on the total number of rows or the total number of
data points.
}
\examples{
# Generate a small matrix
x <- matrix(rnorm(50), nrow = 10)

# Divide based on full rows
train_test_rows <- train_test_split(x, criteria = 'rows', train.pct = 0.5)
x_train <- train_test_rows$x_train
x_test <- train_test_rows$x_test

train_test_points <- train_test_split(x, criteria = 'points', train.pct = 0.5)

}
