% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{train_test_split_points}
\alias{train_test_split_points}
\title{Split points of a given matrix Y into train / test}
\usage{
train_test_split_points(Y, train.pct = NULL, train.size = NULL, seed = NULL)
}
\arguments{
\item{Y}{\eqn{(N \times P)} matrix of predictive variables.}

\item{train.pct}{Float number indicating the \% of rows used for training.}

\item{train.size}{Integer number indicating number of rows used for training. \code{train.size} is superseded by \code{train.pct}.}

\item{seed}{Seed for the random generator number.}
}
\value{
A list containing a matrix Y.train and a matrix Y.test
}
\description{
Split points of a given matrix Y into train / test based on parameters train.pct and train.size. If both are informed train.pct takes preference. This function keeps the dimensions of the original Y in both train and test and substitutes the values in both splits by NAs
}
