% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{train_test_split_rows}
\alias{train_test_split_rows}
\title{Split rows of a given matrix X into train / test}
\usage{
train_test_split_rows(x, train.pct = NULL, train.size = NULL, seed = NULL)
}
\arguments{
\item{x}{N times p matrix of predictive variables.}

\item{train.pct}{Float number indicating the \% of rows used for training}

\item{train.size}{Integer number indicating number of rows used for training. train.size is superseded by train.pct}

\item{seed}{Seed for the random generator number. Parameter included for reproducibility purposes. Default is NULL (meaning no seed is assigned).}
}
\value{
a list containing a matrix x_train and a matrix x_test
}
\description{
Split rows of a given matrix X into train / test based on parameters
train.pct and train.size. If both are informed train.pct takes preference
}
